third_party/WebKit/LayoutTests/imported/wpt/custom-elements/Document-createElement-expected.txt

CONSOLE ERROR: line 225: Uncaught ReferenceError: document_types is not defined
This is a testharness.js-based test.
PASS document.createElement must create an instance of custom elements 
FAIL document.createElement must report a TypeError when the result of Construct is not a DOM node Failed to execute 'createElement' on 'Document': The result must implement HTMLElement interface
FAIL document.createElement must report a TypeError when the result of Construct is a TextNode Failed to execute 'createElement' on 'Document': The result must implement HTMLElement interface
FAIL document.createElement must report a NotSupportedError when attribute is added by setAttribute during construction Failed to execute 'createElement' on 'Document': The result must not have attributes
FAIL document.createElement must report a NotSupportedError when attribute is added by attributes.setNamedItem during construction Failed to execute 'createElement' on 'Document': The result must not have attributes
PASS document.createElement must not report a NotSupportedError when attribute is added and removed during construction 
FAIL document.createElement must report a NotSupportedError when a Text child is added during construction Failed to execute 'createElement' on 'Document': The result must not have children
FAIL document.createElement must report a NotSupportedError when a Comment child is added during construction Failed to execute 'createElement' on 'Document': The result must not have children
FAIL document.createElement must report a NotSupportedError when an element child is added during construction Failed to execute 'createElement' on 'Document': The result must not have children
PASS document.createElement must not report a NotSupportedError when an element child is added and removed during construction 
FAIL document.createElement must report a NotSupportedError when the element gets inserted into another element during construction Failed to execute 'createElement' on 'Document': The result must not have a parent
PASS document.createElement must not report a NotSupportedError when the element is inserted and removed from another element during construction 
Harness: the test ran to completion.
